[REFACTOR] Move invoice mapper service directory
